+++ CHANGELOG.md
@@ -29,6 +29,7 @@
   - **_Breaking Change_**: rename `ParserContext` to `GrammarRuleContext`, make `ParserContext` a class.
   - **_Breaking Change_**: remove `ReducerContext`, just use `GrammarRuleContext`.
   - **_Breaking Change_**: rename `BaseResolverOptions.reduce` to `accept`.
+  - **_Breaking Change_**: rename `ASTNode.toObj` to `toJSON`.
   - Feat: add `ASTNodeSelector`.
   - Feat: resolved conflicts in definition context will have higher priority than resolved conflicts defined in parser builder.
   - Feat: typed parser.